seleccion/cv-extractor/requirements.txt
pdfplumber==0.11.0
anthropic==0.40.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
